remove target, more clean up
